# specs/001-i-want-to/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Lift Tracker API
  description: REST API for workout tracking application
  version: 1.0.0
  contact:
    name: Lift Tracker Team
servers:
  - url: https://api.lifttracker.com/v1
    description: Production server
  - url: https://localhost:7001/api
    description: Development server
security:
  - GoogleOAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Initiate Google OAuth login
      responses:
        '302':
          description: Redirect to Google OAuth
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/callback:
    get:
      tags: [Authentication]
      summary: Handle Google OAuth callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to application with auth token
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/profile:
    get:
      tags: [Authentication]
      summary: Get current user profile
      responses:
        '200':
          description: User profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Workout Session endpoints
...
components:
  securitySchemes:
    GoogleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: Access to email address
            profile: Access to profile information
  schemas:
    UserProfile:
      type: object
      required: [id, email, name, createdDate]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 100
        createdDate:
          type: string
          format: date-time
        lastLoginDate:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...